# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "picweight API",
    "description": "AI-assisted calorie & БЖУ tracker. Photograph what you eat; a server-side estimation agent returns calories and macros.",
    "license": {
      "name": "MIT"
    },
    "version": "master"
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Current user, profile and today's state",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "The current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/me/profile": {
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update body data and recompute targets",
        "description": "Targets are computed with Mifflin-St Jeor — deterministic, offline and auditable. A deficit steeper than ~1% bodyweight/week is returned as a warning, not an error.",
        "operationId": "update_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The saved profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DayState": {
        "type": "object",
        "description": "A day's totals against its targets.\n\nCarried verbatim in the `/meals/events` completion payload so a client\nrenders the notification from one message with no follow-up round trip (§9).",
        "required": [
          "date",
          "consumed_kcal",
          "target_kcal",
          "remaining_kcal",
          "consumed_protein_g",
          "target_protein_g",
          "remaining_protein_g",
          "consumed_fat_g",
          "target_fat_g",
          "consumed_carbs_g",
          "target_carbs_g",
          "meals_logged",
          "status"
        ],
        "properties": {
          "consumed_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate consumed, grams."
          },
          "consumed_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat consumed, grams."
          },
          "consumed_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy consumed, kcal."
          },
          "consumed_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein consumed, grams."
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The local calendar day these figures cover."
          },
          "meals_logged": {
            "type": "integer",
            "format": "int64",
            "description": "Meals logged on this day (all statuses except `failed`)."
          },
          "remaining_kcal": {
            "type": "number",
            "format": "double",
            "description": "`target_kcal - consumed_kcal`; may be negative."
          },
          "remaining_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein still owed, grams; clamped at zero."
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus",
            "description": "Classification derived from the figures above."
          },
          "target_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate allowance, grams."
          },
          "target_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat floor, grams."
          },
          "target_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy target, kcal. Zero when targets are not computed."
          },
          "target_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein floor, grams."
          }
        }
      },
      "DayStatus": {
        "type": "string",
        "description": "How a day is going. Computed from the numbers, never asked of a model.",
        "enum": [
          "on_track",
          "tight",
          "over",
          "protein_unreachable",
          "no_targets"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body returned for every [`AppError`].",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Generic for 5xx, specific for 4xx."
          }
        }
      },
...
      "GoalType": {
        "type": "string",
        "description": "Direction of the calorie goal.",
        "enum": [
          "Lose",
          "Maintain",
          "Gain"
        ]
      },
...
      "MeResponse": {
        "type": "object",
        "description": "`GET /api/v1/me` response.",
        "required": [
          "user",
          "today",
          "version"
        ],
        "properties": {
          "profile": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProfileResponse",
                "description": "Body data and targets. `None` until onboarding completes."
              }
            ]
          },
          "today": {
            "$ref": "#/components/schemas/DayState",
            "description": "Today's state, so the home screen renders from one request."
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse",
            "description": "Identity."
          },
          "version": {
            "type": "string",
            "description": "Backend version string, for the about screen and drift diagnosis."
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "description": "Body data and the targets derived from it.",
        "required": [
          "sex",
          "birth_date",
          "height_cm",
          "activity_factor",
          "goal_type",
          "calibration_factor",
          "timezone"
        ],
        "properties": {
          "activity_factor": {
            "type": "number",
            "format": "double",
            "description": "Activity multiplier, 1.2 sedentary … 1.9 very active."
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "description": "Date of birth."
          },
          "calibration_factor": {
            "type": "number",
            "format": "double",
            "description": "Per-user multiplier learned from correction history (§5 hidden fat)."
          },
          "current_weight_kg": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Most recent logged weight, for convenience."
          },
          "goal_type": {
            "$ref": "#/components/schemas/GoalType",
            "description": "Direction of the goal."
          },
          "height_cm": {
            "type": "number",
            "format": "double",
            "description": "Height in centimetres."
          },
          "rate_kg_per_week": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Intended rate of change, kg/week."
          },
          "sex": {
            "$ref": "#/components/schemas/Sex",
            "description": "Biological sex — changes the BMR constant by 166 kcal."
          },
          "target_carbs_g": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Daily carbohydrate allowance."
          },
          "target_fat_g": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Daily fat floor."
          },
          "target_kcal": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Daily energy target."
          },
          "target_protein_g": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Daily protein floor."
          },
          "target_weight_kg": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Target weight in kilograms."
          },
          "targets_computed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the targets were last recomputed."
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone name, the default for meals that arrive without an offset."
          }
        }
      },
...
      "Sex": {
        "type": "string",
        "description": "Biological sex, as required by the Mifflin-St Jeor BMR formula (§6).",
        "enum": [
          "Male",
          "Female"
        ]
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "description": "Body of `PUT /api/v1/me/profile`.\n\nThis is the onboarding payload and the edit payload both; every field is\nrequired because the formulas need all of them.",
        "required": [
          "sex",
          "birth_date",
          "height_cm",
          "activity_factor",
          "goal_type",
          "target_weight_kg",
          "rate_kg_per_week",
          "timezone"
        ],
        "properties": {
          "activity_factor": {
            "type": "number",
            "format": "double",
            "description": "Activity multiplier, 1.2–1.9."
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "description": "Date of birth."
          },
          "current_weight_kg": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Current weight. When present, also written to `weight_logs` with source\n`onboarding`, since targets recompute on weight change (G6)."
          },
          "goal_type": {
            "$ref": "#/components/schemas/GoalType",
            "description": "Direction of the goal."
          },
          "height_cm": {
            "type": "number",
            "format": "double",
            "description": "Height in centimetres."
          },
          "rate_kg_per_week": {
            "type": "number",
            "format": "double",
            "description": "Intended rate of change, kg/week (always positive)."
          },
          "sex": {
            "$ref": "#/components/schemas/Sex",
            "description": "Biological sex."
          },
          "target_weight_kg": {
            "type": "number",
            "format": "double",
            "description": "Target weight in kilograms."
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone name."
          }
        }
      },
      "UpdateProfileResponse": {
        "type": "object",
        "description": "`PUT /api/v1/me/profile` response.",
        "required": [
          "profile",
          "warnings"
        ],
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse",
            "description": "The saved profile with freshly computed targets."
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Non-fatal advice — e.g. an aggressive deficit. Warned about, not\nsilently accepted (§6)."
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "description": "The authenticated user's identity.",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the account was first seen."
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name from the ID token."
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "Email from the ID token, when the IdP supplied one."
          },
          "id": {
            "type": "string",
            "description": "`users.id`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT. Web clients receive it as the HttpOnly `picweight_session` cookie after OIDC login; native clients obtain it from `POST /api/auth/token` and send it as `Authorization: Bearer`."
      }
    }
  },
...
}